\name{is.posdef}
\alias{is.posdef}
\title{Checks if a matrix is positive definite}
\usage{
is.posdef(matrix, tolerance = 1e-06)
}
\arguments{
  \item{matrix}{input matrix}

  \item{tolerance}{a constant}
}
\value{
A boolean
}
\description{
Checks if any of the eigenvalues of the matrix is smaller
than the set tolerance value
}
\details{

}
\author{
Gaye A.
}
\keyword{internal}

